+++ src/core/transport/index.ts
@@ -4,6 +4,12 @@
 
 export { AdtRequest, hasDeferredResponses } from './AdtRequest';
 export { AdtRequestLegacy } from './AdtRequestLegacy';
+export {
+  addObjectToTransport,
+  createTransportTask,
+  readTransportActionLog,
+  removeObjectFromTransport,
+} from './objects';
 export type { ICreatedTransport } from './parseCreatedTransport';
 export { parseCreatedTransport } from './parseCreatedTransport';
 export { parseTransportTree } from './parseTransportTree';